Move junction angle measurements into standalone CenterlineJunctionAngles module
