+++ CMakeLists.txt
@@ -70,6 +70,7 @@ add_subdirectory(StenosisMeasurement2D)
 add_subdirectory(BranchClipper)
 add_subdirectory(ArterialCalcificationPreProcessor)
 add_subdirectory(CenterlineDisassembly)
+add_subdirectory(CenterlineJunctionAngles)
 add_subdirectory(ClipVessel)
 add_subdirectory(CfdMeshGenerator)
 if(SlicerVMTK_USE_ExtraMarkups)
